fix for https://bugs.eclipse.org/bugs/show_bug.cgi?id=431516
The CreateAppliedStereotypeViewCommand is executed for the creation of a view displaying applied stereotype properties as a comment on diagrams.
The created CSSBasicCompartmentImpl is not displayed because the default css style is used for the visibility of the compartment. 
In order for the CSSBasicCompartmentImpl to be forced to visible the ForceValueHelper.setValue(...) must be used and it is correctly used in the setVisible(...) function of CSSBasicCompartmentImpl but the setVisible function is called too early (not yet a css view) and the ForceValueHelper.setValue() has no effect. 

The setVisible function should be called on the created Node at the end of the CreateAppliedStereotypeViewCommand. 

Change-Id: Ifc8d381356e1135f10783f849243b2bbf673e894
Signed-off-by: Calin GLITIA <calin.glitia@esterel-technologies.com>