458197: [Profile Applications] Customized DSML palettes not enabled in UI when loading Profile Applications https://bugs.eclipse.org/bugs/show_bug.cgi?id=458197
Implement a new Profile Application Listener exactly analogous to the StereotypeElementListener (and, in fact, installed by the same service) that notifies interested parties when profiles are applied to and unapplied from packages in the UML model.  As for stereotype applications in which there is no real feature on which to make the notifications, the custom notifications injected by the ProfileApplicationListener are dispatched via the affected Package and reference an applied or unapplied Profile.

Includes a new test suite for the ProfileApplicationListener (same test scenarios as the previous patch) in the existing test plug-in for oep.uml.tools, alongside the StereotypeElementListener tests.

Change-Id: I1ac73c7fc4aa8b5cec6c0c6abe4be243ba06e9f3