451230: Restore missing edit policy in Use Case Diagram https://bugs.eclipse.org/bugs/show_bug.cgi?id=451230
Restore the Xtend template change lost in a bad merge commit ae20f7c that installs the DefaultGraphicalNodeEditPolicy in NodeEditParts and regenerate the Use Case Diagram.

This fixes a regression in the creation of Associations in use case diagrams (the palette tool failed to create associations between use cases and actors in any combination).