[Releng] Updating dependencies in multiple files
Enhance the dependency updater tool to support
updating dependencies in multiple selected files.  A single
command handler now dispatches to concrete
DependencyUpdaters discovered automatically from the
classpath.

Also added a check to prompt the user when updating
a dependency with a URL that looks quite different, in
case it might be rolling back an integration build to a
previous stable build or something other suspicious
change.  The user is only prompted once per dependency
(the same decision applied to all selected files).