Bug 459566: [Releng] Consider letting Tycho use JDT's project settings for compilation https://bugs.eclipse.org/bugs/show_bug.cgi?id=459566
Add Tycho compiler's <useProjectSettings> to the plug-in management in the main top
POM to ensure that each project's build preferences are respected.

This does require removing a few {@link}/@see references to restricted API in doc
comments to work around a bug in Tycho (459554).

A great many plug-in projects using 1.6 source constructs (mostly @Override of
abstract methods) failed to declare 1.6 requirement either in the BREE, or the
classpath/compiler settings, or both.

Change-Id: Ia73b592fff8d46eaf0bc628b6e21d1e758ff0c64