Bug 456934: Random errors causing org.eclipse.uml2.uml.rsa content type to be disabled https://bugs.eclipse.org/bugs/show_bug.cgi?id=456934
Serialize the determination of content types of files on a single work queue shared
by all WorkspaceModelIndex instances.  This ensures at least that the multiple
concurrent indexing threads don't concurrently initialize EPackages while leaving
the heavy lifting of indexing models (based on EPackages a priori initialized
already on the single worker thread) to be done in parallel jobs.

This also reverts the earlier changes to hack a bootstrapping of package
initialization at start-up before starting the DecoratorModelIndex, which never
actually worked anyways.

Finally, this adds a step that was previously missing in the lifecycle of the
DecoratorModelIndex:  to shut it down when the oep.uml.decoratormodel.ui plug-in is
stopped.