Bug 451230: Modeling assistants customization for profiles https://bugs.eclipse.org/bugs/show_bug.cgi?id=451230
Add a relationshipTypes subset of the elementTypes collection of the assistant
provider to reliably determine which element types are relationships, because
deriving this from the modeled connection assistants would omit element types for
which there is no assistant.

Update the assistants-model generators to reliably generate the relationshipTypes
collection based on (a) links defined in the GMFGen model or (b) UML metaclasses
conforming to Relationship in a profile.

Re-generate the Class and Use Case assistants models.