Bug 459746: Model remains dirty after save and undo https://bugs.eclipse.org/bugs/show_bug.cgi?id=459746
Ensure that, when the undo history is exhausted and we save again, we forget
the save context because now it's as though we're starting afresh with a new
command history.  This lets the isSaveNeeded method rely on the presence of
the save marker to know that we're dirty in the case that there is no more undo.

Includes JUnit tests covering this scenario, the 410310 scenario, and more.

Change-Id: I416fca4fa87c686211f052f9bb148c06179e9efa