Bug 436998: [Control Mode] Exceptions when trying to control multiple elements in the same resource https://bugs.eclipse.org/bugs/show_bug.cgi?id=436998
Implement updates to the CDOAwareModelSet and related tests that were missed when
the base ModelSet behaviour was refactored for control-mode changes.

Basically, override the initial removal of resources from the resource set that are
dirty, because CDO does not permit dirty their removal from the set.  Also, don't
attempt to process cross-unit references on CDOResources because it causes NPEs on
deleted resources later and they don't have cross-references of any kind anyways.

Update the CDO control-mode tests to use the extensible service registry because now
the control-mode framework requires some new services.