Bug 459613: [Profile Applications] [Submodels] Separate Profile Application stereotypes lost for fragmented models https://bugs.eclipse.org/bugs/show_bug.cgi?id=459613
Fix problem of stereotype applications in non-package model units not being moved
into the decorator model resource when a profile application is externalized.

Includes JUnit for that scenario and a new regression test verifying that stereotype
applications pertaining to repeated profile applications in nested packages are not
moved when the application of the same profile at a higher level of nesting is
externalized.