Bug 460392: [Index] Workspace model index job hangs on attempt to cancel https://bugs.eclipse.org/bugs/show_bug.cgi?id=460392
Support cancellation of the index's control job, with automatic re-scheduling on a
one-second delay to let blocked jobs proceed and, in turn, block the index if
necessary.  Also make slightly better use of synchronization primitives to implement
the constraint on the number of parallel project indexing jobs (especially the
incremental re-index follow-ups to a full index and re-tries of failed index jobs).