Bug 460713: [Assistants] Wizard still usable during model generation https://bugs.eclipse.org/bugs/show_bug.cgi?id=460713
Run model transformations in a progress monitor in the wizard. This requires some
new provision for showing useful progress text, which is based on the labels of
the output model element type of each generator.  This, in turn, uses item label
providers supplied by an optional adapter factory injected by the wizard via the
Identifiers object.