Bug 461595: [CSS] Support inference of standard notation styles https://bugs.eclipse.org/bugs/show_bug.cgi?id=461595
Support CSS inference of single-attribute standard styles such as CanonicalStyle.

Fix NPEs resulting from attempts to initialize the CSS engine while a Diagram is
unattached to any resource, such as occurs in the set-up of certain State Machine
Diagram tests that started failing after standard styles such as CanonicalStyle
started delegating to CSS.

Now, the CSS-aware diagram safely creates a local-only CSS engine when not in a
resource and EMF's hook for setting the direct resource is used to purge and
re-initialize the engine when the diagram is added to/removed from a resource.