Bug 461717: [Assistants] Separate semantic and visual element types https://bugs.eclipse.org/bugs/show_bug.cgi?id=461717
Add generation of semantic super-element-type reference in generated visual element
types.  Add an option in the wizard to suppress this semantic supertype reference.

Add JUnit tests to verify the semantic supertype in the diagram-specific scenario and
its absence in the generic element types scenario.

Update the user documentation for the new behaviour and option in the wizard.