Bug 461717: [Assistants] Separate semantic and visual element types https://bugs.eclipse.org/bugs/show_bug.cgi?id=461717
Don't generate redundant apply-stereotypes advice and stereotype matcher for visual
element types when generating reference to the semantic supertype.

Fix a problem in generating the reference to semantic supertype for associations, which
require additional hint suffix on the element type.

Add JUnit tests covering the option to suppress the semantic supertype reference.