[Releng] Papyrus setup model fixes for developer tools installation
Fix the installation of Papyrus Developer Tools by including some missing dependencies,
in particular the nightly Papyrus main build.  Because the developer tools are dependent
on the main Papyrus plug-ins and the setup model installs the latest nightly build of
the Developer tools, it also needs to install the latest nightly build Papyrus, not just
the latest milestone build.