Bug 463564: [Performances] Opening a menu in the ModelExplorer takes several seconds (Mars M6) https://bugs.eclipse.org/bugs/show_bug.cgi?id=463564
Add caching of read-only state for direct queries on resource URIs
(not just via EObjects).

Because many read-only queries are not done within read transactions,
discontinue the tie-in to transactions and, instead, always cache
read-only state and expire the cache (by default) in an async runnable
on the display.

Update the unit tests to inject a custom executor to control when the
cache is cleared.

Change-Id: I04f964a386b3342a1387f98504b0be7bc87de116
Reviewed-on: https://git.eclipse.org/r/44960
Tested-by: Hudson CI
Reviewed-by: Camille Letavernier <camille.letavernier@cea.fr>