Bug 463631: Regression: Referenced models are not read-only https://bugs.eclipse.org/bugs/show_bug.cgi?id=463631
Fixed a bug in the DecoratorModelReadOnlyHandler that caused it to override
lower priority read-only policies with a 'writable' result on resources
that are not decorators of the model being edited.  And a similar treatment
for the user model resources that are decorated by a decorator model opened
as the 'main' model of an editor.

There was also a problem introduced recently in the ReadOnlyCache, which
resulted in the cache only being cleared once.  Subsequently, computed
read-only states of resources would be cached permanently (or, at least,
for the duration of the model set).

Fixing the DecoratorModelReadOnlyHandler bug exposed a latent bug in the
read-only policy for controlled resources.  The ControlledResourceTracker
did not properly handle the topology of a fragmented model in which a
sub-unit has multiple parent units (it stores multiple cross-resource-
contained root elements).  This was detected by JUnit tests for control
mode hanging on a make-writable dialog.

Finally, the ProfileMigrationTest in the decorator-models suite relied on
the broken behaviour of the read-only policy to prevent hanging on the
make-writable dialog when modifying a profile (that should have been
read-only under the standard policy).  This is fixed in the test by
disabling read-only checks in the transaction that modifies the profile.