Bug 466997: [Activity Diagram] Uncaught exception in transaction pre-commit task. https://bugs.eclipse.org/bugs/show_bug.cgi?id=466997
Ensure that diagram refresh runnables execute on the UI thread, forcing submission
of runnables on the fall-back UI executor if necessary when the DiagramHelper is
invoked on some other thread.  Uses a new policy API to let clients like the
DiagramHelper rank the appropriateness of the various executors in the picture.