Bug 465416: Synchronization of diagrams with other diagrams https://bugs.eclipse.org/bugs/show_bug.cgi?id=465416
Fix exceptions on attempt to delete and init-synchronize FinalState because the
nested floating label edit-part is taken as a region.  This is fixed generally by
adding filtering of child edit-parts in the iteration of a parent to process only those
that visualize elements of the semantic child type; and specifically by not adding the
regions sync-features to the sync-buckets for FinalStates in both the semantic and
notation synchronization.