Bug 467207: [Tests] Fix the ClassDiagram ContainmentLink test https://bugs.eclipse.org/bugs/show_bug.cgi?id=467207
Fix test failures caused by the unloading of modeled element-types by the
Modeling Assistants tests, which element-types were retained as pseudo-constants
in the UMLElementTypes classes of generated UML diagrams.

Incidentally fix a problem in the unloading of an element-types configuration
model in which the registry would (a) fail to remove unloaded resources from the
resource set because unloading them disconnects them from the (now proxy)
configuration-set, and (b) unload resources that it hadn't loaded into its own
resource set but are loaded and managed by some client component.