[Tests] Fix test failures
Fix a memory leak in the CDO editor tests, which also happens to be
a failure of the asynchronous refresh throttle mechanism:  the leak occurs when refreshes are never posted by the editor is put into the throttle map,
and when the refreshes are posted they aren't actually throttled.  All of
these problems are now fixed.