Bug 450944: [All diagrams] Block drag&drop of element into the list compartment already containing the same element https://bugs.eclipse.org/bugs/show_bug.cgi?id=450944
Restore drag-and-drop behaviour required by canon edit policy by removing an
unnecessary restriction in the CommonDiagramDragDropEditPolicy.

Annotate a test failure in the Class Diagram tests where nested interface views
in the nested-classifier list compartment are not removed when the interface is
dragged onto the diagram surface (unlike other nested classifiers).