Bug 468079: [Developer] Design documentation generator limitations https://bugs.eclipse.org/bugs/show_bug.cgi?id=468079
Include collaborations in the design model, to generate their behaviour diagrams and
traceability information.

Add traceability cross-references to classifiers in the use case and design models.

Generate references in taxonomy and traceability cross-references as hyperlinks. This
involves a new LinkUtil API for encoding links in the comment text with corresponding
new API in the ITranscription back-end to support transcoding the links to the output
format required by the generated document.