Bug 467016: [State Machine] Cannot create State Machine Diagram in a class https://bugs.eclipse.org/bugs/show_bug.cgi?id=467016
The viewpoint specifies BehavioredClassifier::classifierBehavior as the reference
in which to create the Behavior required by a new behavior diagram of a behaviored
classifier.  That's fine, except that this is not a containment reference, so GMF
rejects it.  The UML-specific ElementEditHelper is updated to accept subsets of
containments also, because the UML2 API implicitly adds elements to the containment
superset, ensuring proper attachment of the new element to the model.