Bug 468071: [Model Explorer] Cannot create redefinable template signature for operations https://bugs.eclipse.org/bugs/show_bug.cgi?id=468071
Operations and other non-Classifier templateable elements cannot own
RedefinableTemplateSignatures because these require the template to be a classifier.
So, add an advice that denies create/move/set commands that attempt to put a
RedefinableTemplateSignature into a non-Classifier.

Add JUnit tests to verify the denial of the forbidden commands and continuing
proper function of commands that make allowable template signature associations.