Bug 469464: [Language Service] Deadlock in the Language Service when starting Papyrus https://bugs.eclipse.org/bugs/show_bug.cgi?id=469464
The workbench creates the user's last opened editors when starting up while the
Job Manager is suspended.  Consequently, if the UI thread blocks waiting for an
index job, it is likely to deadlock and make the user sad.

A fix:

Support asynchronous calculation of model languages in the language service.
Update the Papyrus Editor to schedule a re-load when new languages are added, in case
a language like UML-RT was detected that requires deep model-set configuration.
The possible need to re-load is mitigated by a new synchronous API on the Profile
Index Service that will force (if necessary) a scan of the resource in question for
intrinsically applied profiles, only.  Because we anticipate that UML-RT and other
profiles like it will not actually support externalized application, this should
ensure that the asynchronous completion of the language query should almost never
find additional languages that would trigger a re-load.

Change-Id: Id93d9304a40a272456015aae0b9d054353cc0f67
Reviewed-on: https://git.eclipse.org/r/49526
Reviewed-by: Camille Letavernier <camille.letavernier@cea.fr>
Tested-by: Hudson CI