Bug 473183: [Tests] Improve test performance by sharing diagram editors https://bugs.eclipse.org/bugs/show_bug.cgi?id=473183
Implement a new test rule, employed by the AbstractPapyrusTestCase in
the UML common diagram test API, that creates a single editor shared
by all test cases in the suite class.  The contents of the UML and
notation resources are cleaned up after each test case to restore
them to the initial state as configured by the command that the test
suite uses to create the diagram.

Test-suite classes may be annotated with @StateNotShareable to opt out
of this sharing of diagram editors, if necessary.  This is done for
a handful of test suites that exhibit errors and other failures when
attempting to share a diagram editor.

Change-Id: I216f82b34700b1ad46d8a966294f3b584095a413