Bug 465899: Cannot connect ports of imported library as "Registered package" https://bugs.eclipse.org/bugs/show_bug.cgi?id=465899
Introduce request parameters as hints to the read-only advice's handling
of CreateRelationshipRequest, to tell it when the source/target of the
relationship will or will not be modified by the creation of the
relationship.  Use these hints, when available, to further refine the
determination of read-only violations.

Change the read-only advice implementation for create-relationship
requests to be an after advice, instead of before, to ensure that it
finds as much relevant information in the request as possible that
may contributed by before advices and the edit-helper, itself.

Add, in the UML Element Types, a new advice that sets these hints for
the benefit of improved read-only calculation.  The advice intercepts
creation requests for Associations of all kinds and for Connectors.

JUnit tests are included for association and connector scenarios.
Existing tests cover the correct handling of metaclass extensions.

Change-Id: I9a03fb3ad943a8ae6713a721937bb82a37966562