Bug 473154: FileNotFoundException in LocalFile.openInputStream (369) https://bugs.eclipse.org/bugs/show_bug.cgi?id=473154
Ensure that we only attempt to index files that are synchronized
as far as the workspace is concerned.  Includes a JUnit test that
fails without the fix on detection of the same exception logged
as in the original bug report.
(cherry picked from commit e334737bc2fe21da7ab91bfc87fcd49944275bdd)

Change-Id: Idf211d87b7c59cea73d399419e8dcabede66dde8