Bug 473148: [Synchronization] Classes show attributes for association ends https://bugs.eclipse.org/bugs/show_bug.cgi?id=473148
Ensure that the canonical edit policy only visualizes association ends
inside the attribute compartment in the case that the end's association
will not be visualized.

Includes new JUnit tests for association ends that do and that do not
have their associations presented canonically.  Another existing test needed to be adjusted for this new behaviour.