Bug 476683: [AOF Sync] Investigate/prototype diagram-to-diagram synchronization on AOF https://bugs.eclipse.org/bugs/show_bug.cgi?id=476683
Add API to the JUnitUtils to get all of the annotations applied to a
test case (and/or a test suite).  Add @CleanUp annotation for fields to
be automatically cleaned up by the HouseKeeper rule, that it doesn't
already recognize as needing clean-up.

These changes are Required by the unit test framework for AOF-based
mappings, which consumes the utilities via the Developer Tools build.