Bug 476559: [Table] Sub-optimal performance when expanding subtrees containing lots of children https://bugs.eclipse.org/bugs/show_bug.cgi?id=476559
Manage the list event as a collection of range instead of manage one by
one item.
The validate and invalidate cache is managed only one time by item
expanded (this cache contains only the hierarchical index of the current
element to manage (we can't manage all because the lists will contains
too much elements (and decrease the process time) and because the expand
action managed by nattable allows to expand level by level)).

Change-Id: Ie4b4bc329890fe5cfa610680f49d9f4ed64845d6
Signed-off-by: Nicolas FAUVERGUE <nicolas.fauvergue@all4tec.net>