Bug 476683: [AOF Sync] Support ad hoc tear-down of mappings https://bugs.eclipse.org/bugs/show_bug.cgi?id=476683
An earlier commit implemented automatic undo of the set-up of a
mapping by tracking the observers that were added to observables
and providing a quick means of removing them.  This is not sufficient,
however, for ad hoc removal of a mapping some time after it was
established (ad hoc meaning not as an undo but as a distinct new
action).

This commit refactors the mapping pairs produced by the API as
mapping instances that can be destroyed.  The down-stream changes
that occur as consequences of a mapping are automatically captured
using the injection of observer intercepts using the new AOF core API.