Bug 476683: [AOF Sync] Investigate/prototype diagram-to-diagram synchronization on AOF https://bugs.eclipse.org/bugs/show_bug.cgi?id=476683
Initial prototype of a synchronization service integrating AOF
synchronization mappings into the Papyrus Editor.  Includes a redesign
of the undo/redo support for mappings, discarding the command approach
in favour of modeling the mapping instances and having their attachment
to a model be tracked for undo/redo by EMF together with the user model
content.