Bug 478558: [RobotML] RobotML replaces attribute types and removes connectors when a model is opened https://bugs.eclipse.org/bugs/show_bug.cgi?id=478558
Ensure that the canonical edit policy does not try to visualize the
association of an association-end property in a composite structure
diagram, which causes edit-helpers for SysML to get confused and make
nonsensical changes to the model.

Includes a JUnit regression test based on the attachment to the
original bug report.