Bug 479999: Compilation error in sasheditor on internal API removed in Neon https://bugs.eclipse.org/bugs/show_bug.cgi?id=479999
Replace used of the internal Util API for getting adapters that no longer exists by the corresponding PlatformHelper API in the Papyrus utils bundle.

The sash-editor bundle could not reuse the AdapterUtils from the Papyrus core bundle, and this API was already partially redundant with the earlier API in the lower-level bundle, so a minor refactoring pushes down the AdapterUtils operations to PlatformHelper with compatible signatures.