Bug 480221: [Filters] NPE in the ProfileAppliedFilter when using URIs https://bugs.eclipse.org/bugs/show_bug.cgi?id=480221
Update the ProfileApplied filter's match algorithm to avoid pointless
loading of profiles that aren't the profiles matched by the filter.
This is verified by a new regression test.

This change also exposes errors in tests defined earlier:

* the core ProfileApplied::match test changed the filter's URI to
make the referenced profile unresolvable.  However, the correct
test is to make the mode's actual applied profile unresolvable but
still match the URI expected by the filter

* a test model in the Diagram Assistants tests had a ProfileApplied
filter that was misspecified:  it did not use the correct URI in
referencing the expected profile

Change-Id: If197a1023e9c9ce965f6f48759f76f878211b9cc