Bug 476683: [AOF Sync] Diagram-to-diagram synchronization on AOF https://bugs.eclipse.org/bugs/show_bug.cgi?id=476683
This commit is a squash of several commits on the AOF Synchronization
development branch that fix bugs in the AOF APIs (being problems in the
specification and/or implementation of those APIs, according to their
original intent).

[AOF] Fix NPEs in operations on primitive-valued EAttributes.
commit fcbc4951cdd54e7b8e7ccc2b04d5fd1314d6ad5d

Fix exceptions in CollectBox operations on source box element types that
are abstract, which cannot provide default instances for calculation of
the result box's constraints.

[AOF] The "constants" for empty boxes are mutable
commit 515caf02ce091a3253ffc6d92ea3f34686034a1e

Fix the mutability of the OPTION, ONE, SET, etc. fields of the IBox
interface by introducing actual immutable box implementations and and
replacing these fields of IBox with emptyXyz() operations in a new Boxes
class analogous to java.util.Collections.

Also add utility API for creating immutable boxes from other boxes and
from source elements.  Remove the ConstantOne class that is made
redundant by general support for immutable boxes.

[AOF] More client-friendly observable API
commit 76aefb0551a64c2c074bf363f2ff86cba0b3d74b

Widen the type bounds for observers attached to an observable. Follow
the industry-standard pattern for observers by ignoring attempts to
repeatedly add observers and attempts to remove observers that are not
present, instead of throwing exceptions.

[AOF] List-feature box delegate add/remove notifications broken
commit 47e6b9fe4f0edb2eaa132df1f489ec4df38cabad

Fix list-feature box delegate handling of EMF ADD_MANY and REMOVE_MANY
notifications, which failed to properly account for the insertion and
removal indices respectively and which, in the ADD_MANY case, would also
remove an element because of a missing break statement.

[AOF] Allow redundant collection operations
commit 6701af7a1ce9150259dec50855828d9429cd9c96

It is standard practice for collections in Java to allow redundant add
and remove operations:  adding an already-present element to an unique
collection should simply ignore the request and removing an element that
is not in the collection should similarly be ignored.

[AOF] Fix problems with metaclasses and list features for EMF
commit acce4d4b403bbfab97b3d3c48ba2953b90411e9c

IMetaClass instances from different platforms/domains should not be
comparable by isSubTypeOf(...) but neither should this method throw an
exception.  The result should just always be false for mixed platforms.

The box delegate for EMF list features had two problems (really two
consequences of the same issue):

* the backing-store was not kept up-to-date when changes occurred in
  list from the EMF side, if no client had registered any IObserver

* bindings that should fail catastrophically appeared to work:  direct
  binding of two containment lists, or reference lists with opposites,
  should fail because the lists cannot possibly be made to have the
  same contents

Both of these problems are fixed by ensuring that the adapter that
maintains the backing-store in the ListFeatureDelegate is always active.
And the implementation is slightly streamlined by removing a level of
indirection: delegating directly to a ListDelegate, not to a Box that
delegates to a ListDelegate.

[AOF] Remove obsolete feature delegates from cache
commit 5833e26093ea4f0d4090645a72ea160f1d9966ba

When active operations are removed from EMF objects (as when
uninstalling a synchronization mapping), boxes representing list
features become obsolete but remain in the metaclass's cache. The next
time an active operation is created that operates on the same feature
of the same object, it sees the wrong values and fails silently.

This is fixed by removing the box from the cache when it is no longer
used.

Change-Id: Ia53a498780682678c181abd0fb03384fc4289103