Bug 476683: [AOF Sync] Diagram-to-diagram synchronization on AOF https://bugs.eclipse.org/bugs/show_bug.cgi?id=476683
This commit is a squash of several individual commits implementing
enhancements to the core AOF APIs to support the requirements of the
Papyrus AOF diagram-to-diagram synchronization feature.

[AOF] Support unsettable features and API to query IOne default state
commit d428be27c1a215d0a97adf5732a9828f1d834342

Add support for "unsettable" features in EMF models.  As the IOne API
already provides for the concept of a "default" value, it makes sense
to add API that lets clients query whether the one's value is what it
is because it's defaulted or whether the value was explicitly set but
happens to be the same as the default.

[AOF] Add support for one-way bindings
commit 7b9264b6e89bbd36d187713a58d70cbbf9d9b143

Add API to the IBinding to set the binding to automatically disable
itself on detection of a change to the left-hand side (target of the
right-to-left binding).  This then prevents left-to-right change
propagation and also subsequent overriding of left-hand side by changes
to the right-hand side.  The binding can, of course, be re-enabled at
any time for right-to-left synchronization (and auto-disable can be
turned off at any time, too).

[AOF] Support POJOs in the AOFFactory
commit 2f2ed5a57fd4b81709f3a37b4e6f11da627eb383

The initial implementation of the AOFFactory supported only Java classes
whose properties are natively of IBox type.  To support integration of
POJO types into the AOF, adapters should be provided for properties of
simple scalar and collection types.

[AOF] Refactorings for reuse of test utilities from master branch.
commit 5aeb211f7aae5d5c7bd3945875d1ed8772e8d2a9

[AOF] Support for one-way initial-value bindings
commit 5fec5f2cab427996524d8a3c07299943bb5641d5

Add a new active operation to implement one-way conditional bindings,
in particular for the use case of initial-value bindings.  These are
bindings to a box that fire only when the box has no value yet, to
provide an initial value that wasn't available at the time of creation
of the box.

[AOF] Improving test coverage and tweaks for diagram synchronization
commit 33613fc629122f0301d45eabc3b65fb5cc8e107b

Primarily new APIs are box wrappers for creating multiple distinct
identities for shared boxes and refactoring the protected API of the
EMFMetaClass for extensibility.

[AOF] Automatic observer tracking for clean-up
commit 77c02aa329eb3f4ecfa38d2c46080cd59157857d

Add a facility for tracking the attachment of observers to observables
to later clean them up, so that observed objects may have active
operations easily removed to undo the establishment of those operations.

[AOF] Precise factory tracking
commit e48dc4d759dda523198ca512c3f430e7116c1af1

Boxes and operations need to know the factory that created them
so that they may use the correct factory to create further boxes
and operations.  Also because the factory denotes the platform
or domain of the model elements in the boxes, which clients may
need to interrogate in some box.

[AOF] Add general purpose observer intercept facility
commit 39edae6d666225c796d177c061c04abec8d13aaf

Synchronization mappings need to be able to put certain information
about the active mapping into context whenever the Active Operations
implementing a mapping fire (on detection of changes in the mapping
source and/or target).  To that end, a facility for injecting a
wrapper around observers when they are added to any observable object
is helpful.  This commit adds such a facility and hooks it into the
base observable implementation(s) of the AOF.  A JUnit test is
included.

Also, when adding the new Observers class, it was noticed that the
ObservableTracker utility API was not in the right package, so it
is now moved.

[Releng] Propagate standard settings to new projects
commit c2bbadba3255ff7ae973046f3261f3b9fe242411

[AOF] Design model updates
commit bc01b304fc6189f61cdb5bfb1d3e59372e42adac

Update design model for AOF to use Java package naming convention and
add some new content for new APIs.

[AOF] Expose IBinding enabled state as a box
commit 7fef89597cb0e8b8f1923a39f2f73646a6822849

Encapsulate the enabled state of the IBinding produced by the Bind
operation as a boolean one-box, so that it may be observed and
manipulated by operations.  This provides for a hook by which diagram
synchronization mappings may detect when bindings become disabled by
some user action so that synchronization exceptions may be persisted.

[AOF] Binary cache access API
commit 260b29bf156fed414e4238bdc4328b2d7da7dcdf

Add access to all mappings under the first key of a pair of keys in the
binary cache.

Change-Id: I609322bed8c7a3d2372c0950af78debd79249bfc