Bug 481302: [Profile] when reapply profile you can lose information https://bugs.eclipse.org/bugs/show_bug.cgi?id=481302
Work around the missing "foreign definitions" reference from a dynamic
profile definition to a static definition that it extends. In the
Papyrus implementation of the stereotype-application copier, look up UML
definitions of Ecore elements in the implicitly referenced "foreign
definitions" when not found by the default UML implementation.  This
provides for compatibility with dynamic profiles already deployed that
are missing this annotation (even if an eventual fix for bug 481712 adds
them to profile definitions in the future).

Includes regression tests for:

* dynamic profile extending a dynamic profile (control case)
* dynamic profile extending a static profile, using Papyrus algorithm
* dynamic profile extending a static profile, using UML2 algorithm
(annotated as a failing test)

Change-Id: Ie6830a5bcc1f9246a4476e9160c0ed1ccc00fc6e
Signed-off-by: Christian W. Damus <give.a.damus@gmail.com>