Bug 476683: [AOF Sync] Diagram-to-diagram synchronization on AOF https://bugs.eclipse.org/bugs/show_bug.cgi?id=476683
Add a mechanism to hook in behaviour to be invoked whenever an AOF
binding configured by a mapping is automatically disabled by an override
on the target side.  Employ this mechanism to disable synchronization of
all layout information on a shape whenever some layout detail (position
or size) is overridden, and attach a new style recording the overrides
for subsequent exclusion from synchronization.

This new style is a class in a new model that adds synchronization
styles to the GMF Notation model (as an extending model).

Similarly, implement an annotation for Ecore-based models such as UML2
to persist overrides of synchronization of the semantics underlying
the diagram.