Bug 482927: [Properties] Customization context models should have UI labels https://bugs.eclipse.org/bugs/show_bug.cgi?id=482927
Add an optional 'label' attribute to the Context class and a
derived 'userLabel' that is the 'label' if present, otherwise the
'name'.

Use the new derived user label wherever appropriate in the UI:
in progress monitors, informational dialogs, and especially the
preference pages.  In the "Property view" preference page, sort
contexts by user label.

Update existing context models to provide user-friendly labels.