Add test case for Event_015:
It is possible for more than one Transition to be enabled within a
StateMachine. If that happens, then such Transitions may be in conflict
with each other. For example, consider the case of two Transitions
originating from the same State, triggered by the same event, but with
different guards. If that event occurs and both guard conditions are
true, then at most one of those Transition can fire in a given
run-to-completion step. (p.331)

Change-Id: I273a13deaeda7f88bf15b02aca0f51f91035f49b
Signed-off-by: jeremie.tatibouet <jeremie.tatibouet@cea.fr>