Bug 434983: [SashEditor] Provide a mechanism to move the editor layout model to either the *.di (Legacy) or hidden *.sash resource (1.0) https://bugs.eclipse.org/bugs/show_bug.cgi?id=434983
Implement new actions in the editor's drop-down tool menu for selection
of the the sash-editor page layout storage mode (private or shared).
Also prompt on first opening of a legacy-mode (shared layout) model to
make it private, as determined by the non-existence of a private *.sash
resource, with a corresponding preference page.

Rather than introduce a new service for the purpose, add the handling
of first-open of a model in the SaveBeforeCloseService, which already
has an editor lifecycle listener for managing persistence of the page
layout, and this is another aspect of that persistence management
function.  Don't rename it as SashModelPersistenceService because it
is public API.

Change-Id: I64ec33c6471a16d711426644dba84981e85b2cba