Add test for requirements Exiting_003 and Exiting_005.
Exiting_003: When exiting from a composite State, exit commences with
the innermost State in the active state configuration. This means that
exit Behaviors are executed in sequence starting with the innermost
active State. (p.324)

Exiting_005: When exiting from an orthogonal State, each of its Regions
is exited. After that, the exit Behavior of the State is executed.
(p.324)

Change-Id: If154719684ca8d277fb61e32581a16bf9f34f5d2
Signed-off-by: jeremie.tatibouet <jeremie.tatibouet@cea.fr>